# static/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API to interact with [Wakapi](https://wakapi.dev)\n\n## Authentication\nSet header `Authorization` to your API Key encoded as Base64 and prefixed with `Basic`\n**Example:** `Basic ODY2NDhkNzQtMTljNS00NTJiLWJhMDEtZmIzZWM3MGQ0YzJmCg==`",
        "title": "Wakapi API",
        "contact": {
            "name": "Ferdinand Mütsch",
            "url": "https://github.com/muety",
            "email": "ferdinand@muetsch.io"
        },
        "license": {
            "name": "GPL-3.0",
            "url": "https://github.com/muety/wakapi/blob/master/LICENSE"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/compat/wakatime/v1/users/{user}/heartbeats": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "heartbeat"
                ],
                "summary": "Get heartbeats of user for specified date",
                "operationId": "get-heartbeats",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Date",
                        "name": "date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Username (or current)",
                        "name": "user",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "heartbeat"
                ],
                "summary": "Push a new heartbeat",
                "operationId": "post-heartbeat-3",
                "parameters": [
                    {
                        "description": "A single heartbeat",
                        "name": "heartbeat",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Heartbeat"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Username (or current)",
                        "name": "user",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/compat/wakatime/v1/users/{user}/heartbeats.bulk": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "heartbeat"
                ],
                "summary": "Push new heartbeats",
                "operationId": "post-heartbeat-7",
                "parameters": [
                    {
                        "description": "Multiple heartbeats",
                        "name": "heartbeat",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Heartbeat"
                            }
                        }
                    },
                    {
                        "type": "string",
                        "description": "Username (or current)",
                        "name": "user",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
...
        "/heartbeat": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "heartbeat"
                ],
                "summary": "Push a new heartbeat",
                "operationId": "post-heartbeat",
                "parameters": [
                    {
                        "description": "A single heartbeat",
                        "name": "heartbeat",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Heartbeat"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/heartbeats": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "heartbeat"
                ],
                "summary": "Push new heartbeats",
                "operationId": "post-heartbeat-5",
                "parameters": [
                    {
                        "description": "Multiple heartbeats",
                        "name": "heartbeat",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Heartbeat"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
...
        "/users/{user}/heartbeats": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "heartbeat"
                ],
                "summary": "Push a new heartbeat",
                "operationId": "post-heartbeat-4",
                "parameters": [
                    {
                        "description": "A single heartbeat",
                        "name": "heartbeat",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Heartbeat"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Username (or current)",
                        "name": "user",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/users/{user}/heartbeats.bulk": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "heartbeat"
                ],
                "summary": "Push new heartbeats",
                "operationId": "post-heartbeat-8",
                "parameters": [
                    {
                        "description": "Multiple heartbeats",
                        "name": "heartbeat",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Heartbeat"
                            }
                        }
                    },
                    {
                        "type": "string",
                        "description": "Username (or current)",
                        "name": "user",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
...
        "/v1/users/{user}/heartbeats": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "heartbeat"
                ],
                "summary": "Push a new heartbeat",
                "operationId": "post-heartbeat-2",
                "parameters": [
                    {
                        "description": "A single heartbeat",
                        "name": "heartbeat",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Heartbeat"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Username (or current)",
                        "name": "user",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/v1/users/{user}/heartbeats.bulk": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "heartbeat"
                ],
                "summary": "Push new heartbeats",
                "operationId": "post-heartbeat-6",
                "parameters": [
                    {
                        "description": "Multiple heartbeats",
                        "name": "heartbeat",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Heartbeat"
                            }
                        }
                    },
                    {
                        "type": "string",
                        "description": "Username (or current)",
                        "name": "user",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.Heartbeat": {
            "type": "object",
            "properties": {
                "branch": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "created_at": {
                    "description": "https://gorm.io/docs/conventions.html#CreatedAt",
                    "type": "number"
                },
                "editor": {
                    "description": "ignored because editor might be parsed differently by wakatime",
                    "type": "string"
                },
                "entity": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_write": {
                    "type": "boolean"
                },
                "language": {
                    "type": "string"
                },
                "machine": {
                    "description": "ignored because wakatime api doesn't return machines currently",
                    "type": "string"
                },
                "operating_system": {
                    "description": "ignored because os might be parsed differently by wakatime",
                    "type": "string"
                },
                "project": {
                    "type": "string"
                },
                "time": {
                    "type": "number"
                },
                "type": {
                    "type": "string"
                },
                "user_agent": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}